Missing SetDuty Net Event for Toggling onDuty
When toggling duty on/off, it creates a weird cycle where you are believed to be on/off duty as indicated by the duty check, but the police job resource doesn't see that update. You can't access protected frequencies, vehicle spawns, the armory, etc, without going back to the duty blip multiple times. 

The ToggleDuty event in qb-core calls back to the SetDuty event in the client, and this does not exist in qb-policejob.
https://github.com/qbcore-framework/qb-core/blob/34c574daf295ff1c3f1ebe4447d51a2a82b9ebdc/server/events.lua#L152

I added it along with the check to update the cop count, similar to qb-ambulancejob. 

Fixes #121