Send impounded vehicle correctly
When using the police impound location it is only deleting the vehicle so you need to go to depot to pick up your car.
With this change it sends the car to the police impound and the menu can be used to pickup the vehicle.

Fixes : #339 

Questions (please complete the following information):

Have you personally loaded this code into an updated qbcore project and checked all it's functionality? Yes, tested on multiple servers
Does your code fit the style guidelines? Yes
Does your PR fit the contribution guidelines? Yes